Cleaned up some code

+++ app.rb
@@ -10,7 +10,7 @@ class App < Sinatra::Base
 
     # Elever start page
     get "/elever" do
-        @elever = db.execute("SELECT * FROM elever ORDER BY class, name ASC")
+        @elever = db.execute("SELECT * FROM elever ORDER BY class, name DESC")
         erb :"elever/index"
     end
 
@@ -102,12 +102,16 @@ class App < Sinatra::Base
         elev_class = params["elev_class"]
         elev_image = params["elev_image"]
 
-        File.open("public"+image_save_dir+elev_image["filename"], "w") do | f |
-            File.open(elev_image["tempfile"], "r") do | input |
-                IO.copy_stream(input, f)
+        # Checks if file exists
+        if(!File.exists?("public"+image_save_dir+elev_image["filename"]))
+
+            # Creates a new file at directory and copies tempfile data to new file
+            File.open("public"+image_save_dir+elev_image["filename"], "w") do | f |
+                File.open(elev_image["tempfile"], "r") do | input |
+                    IO.copy_stream(input, f)
+                end
             end
         end
-
 
         sql = "INSERT INTO elever (name, age, description, class, image_url) VALUES(?,?,?,?,?)"
         
@@ -140,15 +144,11 @@ class App < Sinatra::Base
                 file.name = file.name[1..-1]
                 
                 #p file.name
-                info_array = []
-
                 info_array = file.name.split("_")
 
                 info_array[3] = info_array[3].split(".")[0]
 
-
                 info_array.append(image_save_dir+file.name)
-
 
                 elev_sql = "INSERT INTO elever (name, age, class, description, image_url) VALUES(?,?,?,?,?)"
